Corrected invalid read of size
