+++ code/config.mk
@@ -4,8 +4,14 @@ SOCKET					:=	$(addprefix $(SOCKET_DIR), $(SOCKET_FILES))
 
 CONFIGURATION_DIR		:=	configuration/
 CONFIGURATION_FILES		:=	Configuration.cpp	\
-							ServerConfiguration.cpp \
-							Route.cpp \
-							Host.cpp	\
-							parsing.cpp
+							ServerConfiguration.cpp	\
+							Route.cpp	\
+							Host.cpp
 CONFIGURATION			:=	$(addprefix $(CONFIGURATION_DIR), $(CONFIGURATION_FILES))
+
+PARSING_CONF_DIR		:=	configuration/parsing/
+PARSING_CONF_FILES		:=	parsing.cpp	\
+							utils.cpp	\
+							parse_route.cpp	\
+							parse_host.cpp
+PARSING_CONF			:=	$(addprefix $(PARSING_CONF_DIR), $(PARSING_CONF_FILES))
