+++ code/includes/parsing.hpp
@@ -62,6 +62,7 @@ void	parse_server(std::map<ip_t, std::vector<ServerConfiguration> > &conf, std::
 void	skip_line(std::string &file, size_t &i, size_t &line);
 void	skip_wspace(std::string &file, size_t &i, size_t &line);
 short	real_atoi(std::string &file, size_t &i, size_t &line, short max, short len);
+uint8_t	ft_hextoint(std::string &file, size_t &i, size_t &line);
 void	parse_ipv4(std::string &file, size_t &i, size_t &line, std::map<in_addr_t, in_port_t> &ip);
 void	parse_host(std::string &file, size_t &i, size_t &line, ip_t &ip);
 void	parse_ipv6(std::string &file, size_t &i, size_t &line, std::map<ipv6_t, in_port_t> &ip);
